#60 Implemented custom stream filter PollsContentStreamFilter and reusable stream query class PollsStreamQuery.
